Add font policy, rename utils
