+++ utils/generateCsp.js
@@ -70,6 +70,7 @@ var generateCSP = function () {
     addPolicy(Directives.StyleSrc, Policies.UnsafeInline);
     addPolicy(Directives.ImgSrc, Policies.Self);
     addPolicy(Directives.ConnectSrc, Policies.Self);
+    addPolicy(Directives.FontSrc, Policies.Self);
     return Object.entries(policy)
         .map(function (_a) {
         var key = _a[0], value = _a[1];
