+++ utils/generatePermissionsPolicy.js
